Updated links for build status and coverage status
Links had gone to fabianvf repo, and the build status links were no longer functioning.  I guessed that they should be updated to CenterForOpenScience repo.  I tested all and did not update the Code Climate link because it did not work for COS repo but still does for fabianvf - so left that one as is.